型変換

+++ TGS2024/Rock.cpp
@@ -182,7 +182,7 @@ void Rock::FragmentEffect()
 			{
 				// 初速度の設定
 				//v0[i] = 300.0f;
-				v0[i] = rand() % 200 + 300;
+				v0[i] = (float)(rand() % 200) + 300;
 
 				// 発射角度の設定
 				fragment[i].degree = rand() % 45 + 45;
